Fernando dev

+++ src/main/java/com/school/ecommerceupch/services/OrderStatusServiceImpl.java
@@ -3,17 +3,23 @@ package com.school.ecommerceupch.services;
 import com.school.ecommerceupch.controllers.dtos.requests.CreateOrderStatusRequest;
 import com.school.ecommerceupch.controllers.dtos.requests.UpdateOrderStatusRequest;
 import com.school.ecommerceupch.controllers.dtos.responses.BaseResponse;
+import com.school.ecommerceupch.controllers.exceptions.ObjectNotFoundException;
 import com.school.ecommerceupch.entities.OrderStatus;
 import com.school.ecommerceupch.repositories.IOrderStatusRepository;
 import com.school.ecommerceupch.services.interfaces.IOrderStatusService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
+import java.util.Collections;
+
 @Service
 public class OrderStatusServiceImpl implements IOrderStatusService {
-    @Autowired
-    private IOrderStatusRepository repository;
+    private final IOrderStatusRepository repository;
+
+    public OrderStatusServiceImpl(IOrderStatusRepository repository) {
+        this.repository = repository;
+    }
+
     @Override
     public BaseResponse create(CreateOrderStatusRequest request) {
         OrderStatus orderStatus = repository.save(from(request));
@@ -22,7 +28,7 @@ public class OrderStatusServiceImpl implements IOrderStatusService {
                 .data(orderStatus)
                 .message("Order Status created correctly")
                 .success(Boolean.TRUE)
-                .httpStatus(HttpStatus.OK)
+                .httpStatus(HttpStatus.CREATED)
                 .build();
     }
 
@@ -53,20 +59,30 @@ public class OrderStatusServiceImpl implements IOrderStatusService {
 
     @Override
     public BaseResponse delete(Long id) {
+
+        if (!repository.existsById(id))
+            throw new ObjectNotFoundException("Order status not found");
+
         repository.deleteById(id);
 
         return BaseResponse.builder()
-                .data(null)
+                .data(Collections.EMPTY_LIST)
                 .message("Order deleted correctly")
                 .success(Boolean.TRUE)
-                .httpStatus(HttpStatus.OK)
+                .httpStatus(HttpStatus.NO_CONTENT)
                 .build();
     }
 
     @Override
     public OrderStatus findOneAndEnsureExistById(Long id) {
-        return repository.findById(id).orElseThrow(() -> new RuntimeException("Order Status not found"));
+        return repository.findById(id).orElseThrow(() -> new ObjectNotFoundException("Order Status not found"));
     }
+
+    @Override
+    public OrderStatus findOneAndEnsureExistByName(String name) {
+        return repository.findByName(name).orElseThrow(() -> new ObjectNotFoundException("Order Status not found"));
+    }
+
     private OrderStatus from(CreateOrderStatusRequest request) {
         OrderStatus orderStatus = new OrderStatus();
         orderStatus.setName(request.getName());
